Refactor a bit to consolidate packages
